# Lexer for expansions