Replace the regular expression based parsing with a recursive descent parser / lexer. This lets us parse nested variable expansions like `${VAR:-${BLAH:-1}}` and also it avoids the bug we discovered where the current interpolator runs in two passes and can resolve variables more times than expected.