@keithpitt 2017-11-05
Oooh la la. Lexers make my brain asplode... How did you write it?