@lox 2017-11-05
Basically just by hand, with tests to work against. I just pushed up a WIP commit where I've swapped to a recursive descent parser, which is basically a lexer and parser mashed together. It's a small enough format that I couldn't justify a lexer and a parser. Trying to handle nested variables with a stream of tokens was breaking my brain too. Recursion is easier :)