@keithpitt 2017-11-05
Yeah cool, and I've got 1 dumb question :)

What makes this approach better than the regexp one? Arguably the regular expression based solution was a little easier to approach than a lexer. Was there some bug that couldn't be fixed with the old approach?