@lox 2017-11-06
@keithpitt yup, the double interpolation bug we chatted about last week. The current parser runs twice over the same text and it's very difficult to figure out a way to prevent the second parse from reading the output of the first parse. 

The thing is that the grammar we are trying to parse is that it's not really plausible to use a regex for if you want to handle things people expect to work like: `${MY_VAR:-$MY_VAR2}`. It's a non-regular language, so using regex's is gonna be hard.

Have a look at my most recent change. I'd argue it's less complex than the regex version.