@keithpitt 2017-11-06
To be honest I have no idea what a lexer or a scanner is (I've never worked with one before)! So I find it hard to understand how it all works. Maybe some stories in the code on how the data flows and how it's put together?