@lox 2017-11-06
Any ideas on the bug I mentioned or how we'd approach resolving nested interpolation with the current regex model?