@lox 2017-11-06
This is also an awesome post about recursive descent parsing in Ruby: http://weblog.jamisbuck.org/2015/7/30/writing-a-simple-recursive-descent-parser.html

The code in that is awesome.