@keithpitt 2017-11-06
> Any ideas on the bug I mentioned or how we'd approach resolving nested interpolation with the current regex model?

I spent about 5 minutes reviewing the old code and I can't think of a way - I think I see the problem you're describing though.

You're probably right that this is the way to go, it's a tough enough problem that needs something more robust to solve. I'm just always 28% cautious of code that I don't understand and I feel like I won't be able to contribute to. Having this library be external makes it a bit easier since the complexity is hidden away.

As long as there are lots of stories/comments in the code so if someone (like me) reads it, I can follow the data as it goes on it's journey through the scanner and parser then 🚀