@lox 2017-11-06
This is a really basic grammar for the language, I think. I'd love a second set of eyes!

```
EscapedBackslash = "\\"
EscapedDollar    = ( "\$" | "$$")

Text             = { EscapedBackslash | EscapedDollar | all characters - "$" }

Identifier       = letters, { letters | digit }
Expression       = { Text | Expansion }
Expansion        = "$" ( Identifier | Brace )

Brace            = "{" Name [ Identifier | Brace ] "}"
BraceExpression  = { Text - "}" | }

BraceEmptyValue  = ":-" Expression
BraceUnsetValue  = "-" Expression
BraceSubstring   = ":" Number [ ":" Number ]
BraceRequired    = "?" Expression

Body             = { Expression } 
```