@lox 2017-11-07
Ok, this is as clean and simple as I think I can get it. I think it's reasonably readable. 

It fixes the issue with the two passes of variable interpolation causing issues and it handles nested variable expansions and it's 5 times faster than the original implementation.