@lox 2017-11-07
> it's 5 times faster than the original implementation.

But uses +50% more memory. Neither stat is of any use to humans as it was already stupidly fast and uses tiny amounts of memory.