@keithpitt 2017-11-08
> But uses +50% more memory

Like an extra 500 bytes or something? :)