@lox 2017-11-08
```
benchmark                       old ns/op     new ns/op     delta
BenchmarkBasicInterpolate-4     4871          1455          -70.13%

benchmark                       old allocs     new allocs     delta
BenchmarkBasicInterpolate-4     8              12             +50.00%

benchmark                       old bytes     new bytes     delta
BenchmarkBasicInterpolate-4     192           588           +206.25%
```