@lox 2017-11-08
Yeah, `Expansion` is the term that bash uses. I didn't like `Expansion` and `Expression` because my brain reads them the same, but that is what it fell out as, happy to take suggestions!