@keithpitt 2017-11-08
If Bash uses `Expansion`, then that seems like the right term to use!