@lox 2017-11-08
> I got lost in some of the nitty gritties at the bottom of the file, but that's totally OK. I reckon if I needed to work on it, I could. And having it be in 1 file helps a bunch in following the code.

Thanks! Yeah, the fact that it's nibbling forward in lots of 1 unicode rune makes it gnarly.